use float for fps

+++ stable_baselines3/common/off_policy_algorithm.py
@@ -432,7 +432,7 @@ class OffPolicyAlgorithm(BaseAlgorithm):
         Write log.
         """
         time_elapsed = max((time.time_ns() - self.start_time) / 1e9, sys.float_info.epsilon)
-        fps = int((self.num_timesteps - self._num_timesteps_at_start) / time_elapsed)
+        fps = (self.num_timesteps - self._num_timesteps_at_start) / time_elapsed
         self.logger.record("time/episodes", self._episode_num, exclude="tensorboard")
         if len(self.ep_info_buffer) > 0 and len(self.ep_info_buffer[0]) > 0:
             self.logger.record("rollout/ep_rew_mean", safe_mean([ep_info["r"] for ep_info in self.ep_info_buffer]))
